Begin using C++Tk.
C++Tk submits strings to the Tcl interpreter. It allows C++ code to be
written in such a way that it looks like Tk commands.

As a first step, submit the whole Git GUI code to the Tcl interpreter. It
can't do a whole lot, though, because it cannot find variable argv0. But
we have proof that the library works.

The user must specify the path to C++Tk on the cmake command line using

   -DCPPTK_SOURCE_DIR=/path/to/C++Tk