Mirror whether themed widgets are used in C++.
Note that gui.usettk is an option that can only be set from in the Git
configuration, but not in the Options dialog. For this reason, it need not
be treated in every apply_config call, because the option cannot change
while the program is running.

Move the treatment of usettk into a separate function and call it from c++.

This is needed because we have to synchronize C++'s useTtk() and Tk's
use_ttk, so that the two languages agree on whether panedwindow's sash or
sashpos options must be used.