Move bind_button3() implementation to class GitGui.
We need it for the context menus in the main window.

Ideally, we would just write the template function into gitgui.h. But that
would require to #include <cpptk.h> in that header. We want to avoid that
if possible. As we will need only two different instances, write out the
signatures as regular functions and use the templated function only in the
implementation.